restore ts::table_collection as a valid type name
