+++ examples/confirm_mutation_counts.hpp
@@ -6,13 +6,13 @@
 #include <vector>
 #include <cstdint>
 #include <fwdpp/internal/sample_diploid_helpers.hpp>
-#include <fwdpp/ts/std_table_collection.hpp>
+#include <fwdpp/ts/table_collection.hpp>
 
 #ifndef NDEBUG
 template <typename poptype>
 void
 confirm_mutation_counts(poptype &pop,
-                        const fwdpp::ts::std_table_collection &tables)
+                        const fwdpp::ts::table_collection &tables)
 {
     std::vector<std::size_t> keys;
     for (auto &mr : tables.mutations)
@@ -53,7 +53,7 @@ confirm_mutation_counts(poptype &pop,
 #else
 template <typename poptype>
 void
-confirm_mutation_counts(poptype &, const fwdpp::ts::std_table_collection &)
+confirm_mutation_counts(poptype &, const fwdpp::ts::table_collection &)
 {
 }
 #endif
