+++ examples/simplify_tables.hpp
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 #include <vector>
-#include <fwdpp/ts/std_table_collection.hpp>
+#include <fwdpp/ts/table_collection.hpp>
 #include <fwdpp/ts/table_collection_functions.hpp>
 #include <fwdpp/ts/table_simplifier.hpp>
 #include <fwdpp/ts/count_mutations.hpp>
@@ -17,7 +17,7 @@ template <typename Simplifier, typename poptype>
 std::pair<std::vector<fwdpp::ts::table_index_t>, std::vector<std::size_t>>
 simplify_tables(poptype &pop, const fwdpp::uint_t generation,
                 std::vector<fwdpp::uint_t> &mcounts_from_preserved_nodes,
-                fwdpp::ts::std_table_collection &tables, Simplifier &simplifier,
+                fwdpp::ts::table_collection &tables, Simplifier &simplifier,
                 const fwdpp::ts::table_index_t first_sample_node,
                 const std::size_t num_samples,
                 std::vector<fwdpp::ts::table_index_t> &preserved_nodes,
@@ -61,11 +61,13 @@ simplify_tables(poptype &pop, const fwdpp::uint_t generation,
             confirm_mutation_counts(pop, tables);
             fwdpp::ts::remove_fixations_from_haploid_genomes(
                 pop.haploid_genomes, pop.mutations, pop.mcounts,
-                mcounts_from_preserved_nodes, 2 * static_cast<fwdpp::ts::table_index_t>(pop.diploids.size()), false);
+                mcounts_from_preserved_nodes,
+                2 * static_cast<fwdpp::ts::table_index_t>(pop.diploids.size()), false);
 
             fwdpp::ts::flag_mutations_for_recycling(
-                pop, mcounts_from_preserved_nodes, 2 * static_cast<fwdpp::ts::table_index_t>(pop.diploids.size()), generation,
-                std::false_type(), std::false_type());
+                pop, mcounts_from_preserved_nodes,
+                2 * static_cast<fwdpp::ts::table_index_t>(pop.diploids.size()),
+                generation, std::false_type(), std::false_type());
             confirm_mutation_counts(pop, tables);
         }
     else // Need to remove non-preserved variants from the hash table
