+++ fwdpp/ts/detail/Makefile.am
@@ -1,4 +1,4 @@
 pkgincludedir=$(prefix)/include/fwdpp/ts/detail
 
-pkginclude_HEADERS= advance_marginal_tree_policies.hpp \
+pkginclude_HEADERS= table_collection.hpp advance_marginal_tree_policies.hpp \
 	generate_data_matrix_details.hpp
