+++ testsuite/tree_sequences/empty_table_collection.hpp
@@ -1,11 +1,11 @@
 #ifndef FWDPP_TESTSUITE_EMPTY_TABLE_COLLECTION_HPP
 #define FWDPP_TESTSUITE_EMPTY_TABLE_COLLECTION_HPP
 
-#include <fwdpp/ts/std_table_collection.hpp>
+#include <fwdpp/ts/table_collection.hpp>
 
 struct empty_table_collection
 {
-    fwdpp::ts::std_table_collection tables;
+    fwdpp::ts::table_collection tables;
     std::vector<fwdpp::ts::table_index_t> empty_samples;
 
     empty_table_collection() : tables(1.), empty_samples{} {}
