+++ testsuite/tree_sequences/simple_table_collection_polytomy.hpp
@@ -1,7 +1,7 @@
 #ifndef FWDPP_TESTSUITE_SIMPLE_TABLE_COLLECTION_POLYTOMY_HPP
 #define FWDPP_TESTSUITE_SIMPLE_TABLE_COLLECTION_POLYTOMY_HPP
 
-#include <fwdpp/ts/std_table_collection.hpp>
+#include <fwdpp/ts/table_collection.hpp>
 #include <fwdpp/ts/table_collection_functions.hpp>
 #include <fwdpp/ts/tree_visitor.hpp>
 
@@ -16,10 +16,10 @@ class simple_table_collection_polytomy
 //	   0 1 2  3 4
 {
   private:
-    fwdpp::ts::std_table_collection
+    fwdpp::ts::table_collection
     init_tables()
     {
-        fwdpp::ts::std_table_collection t(1.);
+        fwdpp::ts::table_collection t(1.);
         t.push_back_node(3, 0);
         t.push_back_node(3, 0);
         t.push_back_node(3, 0);
@@ -42,11 +42,11 @@ class simple_table_collection_polytomy
     }
 
   public:
-    fwdpp::ts::std_table_collection tables;
+    fwdpp::ts::table_collection tables;
     std::vector<fwdpp::ts::table_index_t> samples;
     // NOTE: tv is auto advanced to the first tree
     // by the constructor
-    fwdpp::ts::tree_visitor<fwdpp::ts::std_table_collection> tv;
+    fwdpp::ts::tree_visitor<fwdpp::ts::table_collection> tv;
     double total_time;
     explicit simple_table_collection_polytomy()
         : tables(init_tables()), samples{{0, 1, 2, 3, 4}},
@@ -58,7 +58,7 @@ class simple_table_collection_polytomy
     void
     reset_visitor(bool update_samples_list)
     {
-        tv = fwdpp::ts::tree_visitor<fwdpp::ts::std_table_collection>(
+        tv = fwdpp::ts::tree_visitor<fwdpp::ts::table_collection>(
             tables, samples, fwdpp::ts::update_samples_list(update_samples_list));
         tv();
     }
@@ -69,7 +69,7 @@ class simple_table_collection_polytomy
         bool update_samples_list)
     {
         samples = new_samples_list;
-        tv = fwdpp::ts::tree_visitor<fwdpp::ts::std_table_collection>(
+        tv = fwdpp::ts::tree_visitor<fwdpp::ts::table_collection>(
             tables, new_samples_list,
             fwdpp::ts::update_samples_list(update_samples_list));
         tv();
